refactor: put all the logic from MainLayout to SearchProvider
